+++ BoardGameGeek/Extensions/Alert+showError.swift
@@ -1,0 +1,21 @@
+//
+//  Alert+showError.swift
+//  BoardGameGeek
+//
+//  Created by Nafisa Rahman on 30/1/2023.
+//
+
+import UIKit
+
+extension UIViewController {
+
+    func showError(_ error: Error) {
+        let alertController = UIAlertController(title: "Error", message: error.localizedDescription, preferredStyle: .alert)
+        let alertAction = UIAlertAction(title: "OK", style: .default) { [unowned self] _ in
+            self.dismiss(animated: true, completion: nil)
+        }
+        alertController.addAction(alertAction)
+        present(alertController, animated: true, completion: nil)
+    }
+
+}
