+++ BoardGameGeek/View/SearchList/BoardGameController.swift
@@ -83,15 +83,6 @@ private extension BoardGameViewController {
         navigationItem.preferredSearchBarPlacement = .stacked
     }
 
-    func showError(_ error: Error) {
-        let alertController = UIAlertController(title: "Error", message: error.localizedDescription, preferredStyle: .alert)
-        let alertAction = UIAlertAction(title: "OK", style: .default) { [unowned self] _ in
-            self.dismiss(animated: true, completion: nil)
-        }
-        alertController.addAction(alertAction)
-        present(alertController, animated: true, completion: nil)
-    }
-
     func evaluateState() {
         switch boardGameViewModel.state {
         case .loading:
@@ -140,7 +131,7 @@ extension BoardGameViewController: UISearchBarDelegate {
                 self?.searchController.searchBar.text = ""
                 self?.activityIndicatorView.stopAnimating()
                 self?.tableView.isUserInteractionEnabled = true
-                showError(error)
+                self?.showError(error)
             }
         }
